requirements-voice.txt
# Additional dependencies for the Retell AI voice agent integration.
# Install alongside the base requirements:
#   pip install -r requirements.txt -r requirements-voice.txt

retell-ai>=4.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
